refactor: move app store info into app folder
