+++ src/common/components/pages/UserDefinedSpace.tsx
@@ -1,11 +1,11 @@
 import { indexOf, isNil, mapValues, noop } from "lodash";
 import React, { useCallback, useEffect, useMemo, useState } from "react";
 import { useAuthenticatorManager } from "@/authenticators/AuthenticatorManager";
-import { useAppStore } from "@/common/data/stores";
+import { useAppStore } from "@/common/data/stores/app";
 import createIntialPersonSpaceConfigForFid from "@/constants/initialPersonSpace";
 import SpaceWithLoader from "../templates/SpaceWithLoader";
 import { SpaceConfig } from "../templates/Space";
-import { UpdatableSpaceConfig } from "@/common/data/stores/space/spaceStore";
+import { UpdatableSpaceConfig } from "@/common/data/stores/app/space/spaceStore";
 import Profile from "@/fidgets/ui/profile";
 
 const FARCASTER_NOUNSPACE_AUTHENTICATOR_NAME = "farcaster:nounspace";
